_openslide_read_key_file(): Limit max_size
to guard against drivers passing an unreasonable value.

+++ src/openslide-private.h
@@ -143,7 +143,7 @@ void _openslide_int64_free(gpointer data);
 
 /* g_key_file_load_from_file wrapper */
 bool _openslide_read_key_file(GKeyFile *key_file, const char *filename,
-                              int64_t max_size, GKeyFileFlags flags,
+                              int32_t max_size, GKeyFileFlags flags,
                               GError **err);
 
 /* fopen() wrapper which properly sets FD_CLOEXEC */
